requirements.txt
dnspython==2.6.1
pyaes==1.6.1
pymongo==4.6.2
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
